+++ framework-crates/objc2-pass-kit/src/lib.rs
@@ -1,0 +1,27 @@
+//! # Bindings to the `PassKit` framework
+//!
+//! See [Apple's docs][apple-doc] and [the general docs on framework crates][framework-crates] for more information.
+//!
+//! [apple-doc]: https://developer.apple.com/documentation/passkit/
+//! [framework-crates]: https://docs.rs/objc2/latest/objc2/topics/about_generated/index.html
+#![no_std]
+#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+// Update in Cargo.toml as well.
+#![doc(html_root_url = "https://docs.rs/objc2-pass-kit/0.3.0")]
+
+#[cfg(feature = "alloc")]
+extern crate alloc;
+
+#[cfg(feature = "std")]
+extern crate std;
+
+mod generated;
+#[allow(unused_imports, unreachable_pub)]
+pub use self::generated::*;
+
+#[allow(non_upper_case_globals)]
+#[cfg(feature = "PKPass_Types")]
+impl PKPassType {
+    #[doc(alias = "PKPassTypeAny")]
+    pub const Any: Self = Self(!0);
+}